% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Argument-class}
\alias{Argument}
\alias{Argument-class}
\title{Argument}
\description{
Argument
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{value}}{ANY}

\item{\code{type}}{character}

\item{\code{default}}{ANY}

\item{\code{help}}{character}
}}
\author{
Xiaobei Zhao
}

