% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.loglevel}
\alias{as.loglevel}
\title{Coerces an object to loglevel}
\usage{
as.loglevel(x, loglevels = get.loglevel())
}
\arguments{
\item{x}{object}

\item{loglevels}{the defined loglevels}
}
\value{
loglevel
}
\description{
Coerces an object to \code{loglevel}
}
\author{
Xiaobei Zhao
}

