% Generated by roxygen2 (4.0.1): do not edit by hand
\name{atos}
\alias{atos}
\title{Convert an R object to a string}
\usage{
atos(x, envir = sys.frame(sys.parent(0)))
}
\arguments{
\item{x}{an R object.}

\item{envir}{the \code{environment} to use.}
}
\value{
character
}
\description{
Convert an R object to a string
}
\author{
Xiaobei Zhao
}

