% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cat0}
\alias{cat0}
\title{Cat without space but with a newline at the end}
\usage{
cat0(..., file = "", sep = "", fill = FALSE, labels = NULL,
  append = FALSE)
}
\arguments{
\item{...}{see \code{cat}}

\item{file}{see \code{cat}}

\item{sep}{see \code{cat}}

\item{fill}{see \code{cat}}

\item{labels}{see \code{cat}}

\item{append}{see \code{cat}}
}
\value{
see \code{cat}
}
\description{
Cat without space but with a newline at the end by default
}
\author{
Xiaobei Zhao
}

