% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfchunk}
\alias{dfchunk}
\title{Chunk data.frame into parts}
\usage{
dfchunk(x, n, balance.size = TRUE, balance.order = FALSE)
}
\arguments{
\item{x}{data.frame or matrix}

\item{n}{numeric, the number of chunks}

\item{balance.size}{logical, see \code{vchunk}}

\item{balance.order}{logical, see \code{vchunk}}
}
\value{
a list of data.frame
}
\description{
Chunk data.frame into parts
}
\examples{
dfchunk(iris,n=5)
dfchunk(iris[1:20,],n=3)
dfchunk(iris[1:20,],n=3,balance.order=TRUE)
}
\author{
Xiaobei Zhao
}

