% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfconcat}
\alias{dfconcat}
\title{Concatenate data.frame into a string}
\usage{
dfconcat(x, sep = " ", ...)
}
\arguments{
\item{x}{data.frame or matrix}

\item{sep}{character, a delimiter}

\item{...}{further arguments passed to `format`. See \code{format}}
}
\value{
data.frame
}
\description{
Concatenate data.frame into a string
}
\author{
Xiaobei Zhao
}

