% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfsort}
\alias{dfsort}
\title{Sort data.frame given levels of one column}
\usage{
dfsort(x, which.col, levels)
}
\arguments{
\item{x}{data.frame}

\item{which.col}{column index or name}

\item{levels}{character see \code{base::factor}}
}
\value{
data.frame
}
\description{
Sort data.frame given levels of one column
}
\examples{
data(CO2)
dfsort(CO2,"Treatment",c("nonchilled","chilled"))
dfsort(CO2,3,c("chilled","nonchilled"))
}
\author{
Xiaobei Zhao
}

