% Generated by roxygen2 (4.0.1): do not edit by hand
\name{endswith}
\alias{endswith}
\title{Determine if a character string "ends with" specified characters}
\usage{
endswith(x, char, ignore.case = FALSE)
}
\arguments{
\item{x}{character, a string}

\item{char}{character to match}

\item{ignore.case}{logical, whether case is ignored}
}
\value{
logical
}
\description{
Determine if a character string "ends with" specified characters
}
\examples{
endswith('Hello World','world',ignore.case=TRUE)
}
\author{
Xiaobei Zhao
}

