% Generated by roxygen2 (4.0.1): do not edit by hand
\name{func}
\alias{func}
\title{Funciton with attributes}
\usage{
func(x, name, package)
}
\arguments{
\item{x}{function}

\item{name}{character, the function name}

\item{package}{character, where the function is from}
}
\value{
named funciton
}
\description{
Funciton with attributes (name, package)
}
\examples{
\dontrun{
func(lm,'lm','stats')
}
}
\author{
Xiaobei Zhao
}

