% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_executable}
\alias{get_executable}
\title{Get the executable file path of a package}
\usage{
get_executable(pkg, name = tolower(pkg), dir = "bin", mustWork = TRUE)
}
\arguments{
\item{pkg}{character, the name of a package}

\item{name}{character, the name of the executable file}

\item{dir}{character, the directory in the package hierarchy}

\item{mustWork}{See \code{system.file}}
}
\value{
character, the executable file path
}
\description{
Get the executable file path of a package
}
\examples{
\dontrun{
try(get_executable('Xmisc','Xmisc-argumentparser.R'))
}
}
\author{
Xiaobei Zhao
}

