% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getone,list-method}
\alias{getone,list-method}
\title{Get an element by index or name from a list}
\usage{
\S4method{getone}{list}(obj, x, safe = TRUE, msg = TRUE)
}
\arguments{
\item{obj}{list}

\item{x}{index or name of an element}

\item{safe}{whether allow to get an element by index,
in case this element is after any named element}

\item{msg}{whether print a message}
}
\value{
list
}
\description{
Get an element by index or name from a list
}
\examples{
ll <- list(11,22,33,a=44,b=55,66,77,c=88,99)
getone(ll,numeric())
getone(ll,"a")
getone(ll,"c")
getone(ll,1)
getone(ll,7)
getone(ll,7,safe=FALSE)
}
\author{
Xiaobei Zhao
}

