% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.activeBindingFunction}
\alias{is.activeBindingFunction}
\title{General test of a class name being activeBindingFunction}
\usage{
is.activeBindingFunction(cls)
}
\arguments{
\item{cls}{the name of a class}
}
\value{
logical
}
\description{
General test of a class name being activeBindingFunction
}
\author{
Xiaobei Zhao
}

