% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.connection}
\alias{is.connection}
\title{Is a connection}
\usage{
is.connection(x)
}
\arguments{
\item{x}{R object}
}
\value{
logical
}
\description{
Is a connection
}
\examples{
is.connection(textConnection(LETTERS))
}
\author{
Xiaobei Zhao
}

