% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.loglevel}
\alias{is.loglevel}
\title{General test of an object being interpretable as loglevel}
\usage{
is.loglevel(x, loglevels = get_loglevel())
}
\arguments{
\item{x}{object}

\item{loglevels}{the defined loglevels}
}
\value{
logical
}
\description{
General test of an object being interpretable as \code{loglevel}
}
\author{
Xiaobei Zhao
}

