% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.uninitializedField}
\alias{is.uninitializedField}
\title{General test of a class being uninitializedField}
\usage{
is.uninitializedField(x)
}
\arguments{
\item{x}{class}
}
\value{
logical
}
\description{
General test of a class being uninitializedField
}
\author{
Xiaobei Zhao
}
\seealso{
\code{methods::ReferenceClasses}
}

