% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{List-class}
\alias{List}
\alias{List-class}
\title{A class inherited directly from envRefClass}
\description{
A class inherited directly from envRefClass
}
\section{Fields}{

\describe{
\item{\code{data}}{list, a base::list}
}}
\section{Methods}{

\describe{
\item{\code{popmany(x)}}{Pop many by indexes.}

\item{\code{popone(x, warn = TRUE, error = TRUE)}}{Pop the one at the given index/position (or name) in the list, and return it. If no index is specified, obj$popone() removes and returns the last one in the list.}

\item{\code{removeone(x)}}{Remove the first matched element whose value is x. Display an error if it does not exist.}
}}
\author{
Xiaobei Zhao
}

