% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logsave}
\alias{logsave}
\title{Log a `save`}
\usage{
logsave(x, logger = NULL, envir = sys.frame(sys.parent(0)))
}
\arguments{
\item{x}{ANY, an R object.}

\item{logger}{see \code{\link{logme}}}

\item{envir}{the \code{environment} to use.}
}
\description{
Log a `save`
}
\examples{
inFpath <- "mydir/mypath"
logsave(inFpath)
}
\author{
Xiaobei Zhao
}

