% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lstrip}
\alias{lstrip}
\title{Strip a string with given characters at the beginning (left end)}
\usage{
lstrip(x, char = " ")
}
\arguments{
\item{x}{character, a string.}

\item{char}{character to trim.}
}
\value{
character
}
\description{
Strip a string with given characters at the beginning (left end)
}
\author{
Xiaobei Zhao
}

