% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.dir}
\alias{make.dir}
\title{Make a directory recursively}
\usage{
make.dir(x, mode)
}
\arguments{
\item{x}{character, a directory name.}

\item{mode}{the mode of the path, see \code{dir.create}}
}
\description{
Make a directory recursively
}
\examples{
try(make.dir('~/tmpdir','751'))
}
\author{
Xiaobei Zhao
}

