% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{popmany,list-method}
\alias{popmany,list-method}
\title{Remove and return many elements from a list}
\usage{
\S4method{popmany}{list}(obj, x)
}
\arguments{
\item{obj}{list}

\item{x}{indexes}
}
\value{
the elements and the R object is altered in place
}
\description{
Remove and return many elements from a list
}
\author{
Xiaobei Zhao
}

