% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{popmany-methods}
\alias{popmany}
\alias{popmany-methods}
\title{popmany-methods}
\usage{
popmany(obj, ..., envir)
}
\arguments{
\item{obj}{object}

\item{...}{further arguments}

\item{envir}{the \code{environment} to use}
}
\description{
popmany-methods
}

