% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{popone,list-method}
\alias{popone,list-method}
\title{Remove and return an element from a list}
\usage{
\S4method{popone}{list}(obj, x, warn = TRUE, error = TRUE)
}
\arguments{
\item{obj}{an R object}

\item{x}{name or index}

\item{warn}{logical, whether warn at error}

\item{error}{logical, whether stop at error}
}
\value{
the element and the R object is altered in place
}
\description{
Remove and return an element from an R object, given name or index.
Default: the last element
Note: the R object is altered in place, like Python List pop()
}
\examples{
ll <- list(1,2,3,a=4,b=5,6,7,c=8,9)
popone(ll,"a")
popone(ll,1) ## remove the 1st in ll
popone(ll,1) ## remove the next (2nd of the origin)
try(popone(list(),"a"))
}
\author{
Xiaobei Zhao
}

