% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{popone-methods}
\alias{popone}
\alias{popone-methods}
\title{popone-methods}
\usage{
popone(obj, ..., envir)
}
\arguments{
\item{obj}{object}

\item{...}{further arguments}

\item{envir}{the \code{environment} to use}
}
\description{
popone-methods
}

