% Generated by roxygen2 (4.0.1): do not edit by hand
\name{printme}
\alias{printme}
\title{Print the name and the content of an R object}
\usage{
printme(x = NULL, prefix = NULL, envir = sys.frame(sys.parent(0)))
}
\arguments{
\item{x}{ANY, an R object.}

\item{prefix}{the prefix to print.}

\item{envir}{the \code{environment} to use.}
}
\description{
Print the name and the content of an R object
}
\examples{
## print an object
x1 <- 1:6
printme(x1)

## print with a prefix
foo <- function(x,envir=sys.frame(sys.parent(0))){
  printme(x,match.call(),envir=envir)
  invisible()
}
foo(1:6)
}
\author{
Xiaobei Zhao
}
\seealso{
\code{\link{logme}}
}

