% Generated by roxygen2 (4.0.1): do not edit by hand
\name{R5.value.default}
\alias{R5.value.default}
\title{R5.value.default}
\usage{
R5.value.default(type, default = "list")
}
\arguments{
\item{type}{the type of an R object}

\item{default}{the default value}
}
\value{
ANY
}
\description{
R5.value.default
}
\examples{
require(Xmisc)
R5.value.default('character')
try(R5.value.default(NULL))
R5.value.default('environment')

R5.value.default('hclust')
R5.value.default('dendrogram')
R5.value.default('formula')
R5.value.default('lm')
}
\author{
Xiaobei Zhao
}

