% Generated by roxygen2 (4.0.1): do not edit by hand
\name{R5.value.parse}
\alias{R5.value.parse}
\title{R5.value.parse}
\usage{
R5.value.parse(value, type, default = "list")
}
\arguments{
\item{value}{the value passed to the parameter}

\item{type}{the type of an R object}

\item{default}{the default value}
}
\value{
ANY
}
\description{
R5.value.parse
}
\examples{
R5.value.parse(NULL,'logical')
R5.value.parse(1,'logical')

R5.value.parse(NULL,'hclust')
R5.value.parse(NULL,'dendrogram')
R5.value.parse(NULL,'formula')
R5.value.parse(NULL,'lm')

R5.value.parse(NULL,'character')
R5.value.parse("",'character')
## [1] ""
}
\author{
Xiaobei Zhao
}

