% Generated by roxygen2 (4.0.1): do not edit by hand
\name{raw_input}
\alias{raw_input}
\title{Input from the terminal (in interactive use)}
\usage{
raw_input(msg = "", choice, strip = TRUE)
}
\arguments{
\item{msg}{character, a message to input}

\item{choice}{character, choices to confine the input}

\item{strip}{logical, whether to strip trailing spaces}
}
\value{
character
}
\description{
Input from the terminal (in interactive use),
confined by \code{choice} if provided.
}
\examples{
\dontrun{
raw_input("Please enter user name: ")
raw_input("Please confirm",choice=c("yes","no"))
}
}
\author{
Xiaobei Zhao
}

