% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{removeone,list-method}
\alias{removeone,list-method}
\title{Remove an element from an R object}
\usage{
\S4method{removeone}{list}(obj, x, warn = TRUE, error = TRUE)
}
\arguments{
\item{obj}{an R object}

\item{x}{an element}

\item{warn}{logical, whether warn at error}

\item{error}{logical, whether stop at error}
}
\value{
NULL and the R object is altered in place
}
\description{
Remove an element from an R object
Note: the R object is altered in place
}
\examples{
ll=list(1,2,3,a=4,b=5,6,7,c=8,9)
removeone(ll,3)
}
\author{
Xiaobei Zhao
}

