% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{removeone-methods}
\alias{removeone}
\alias{removeone-methods}
\title{removeone-methods}
\usage{
removeone(obj, ..., envir = sys.frame(sys.parent(2)))
}
\arguments{
\item{obj}{object}

\item{...}{further arguments}

\item{envir}{the \code{environment} to use}
}
\description{
removeone-methods
}

