% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rstrip}
\alias{rstrip}
\title{Strip a string with given chars at the (right) end}
\usage{
rstrip(x, char = " ")
}
\arguments{
\item{x}{character, a string.}

\item{char}{character to trim.}
}
\value{
character
}
\description{
Strip a string with given chars at the (right) end
}
\author{
Xiaobei Zhao
}

