% Generated by roxygen2 (4.0.1): do not edit by hand
\name{schunk}
\alias{schunk}
\title{Chunk a string into parts}
\usage{
schunk(x, size, brk = "-", concat = TRUE)
}
\arguments{
\item{x}{character, a string to chunk.}

\item{size}{numeric, the size of a chunk.}

\item{brk}{character to link broken words.}

\item{concat}{logical, whether to concatenate by a `newline`}
}
\value{
character
}
\description{
Chunk a string into parts
}
\examples{
x <- 'The quick brown fox jumps over the lazy dog.'
cat(schunk(x,15),'\\n')
x <- 'The word, honorificabilitudinita, occurs in Shakespeare\\'s
play Love\\'s Labour\\'s Lost, and means "with honorablenesses".'
cat(schunk(x,30),'\\n')
## The word, honorificabilitudini-
## ta, occurs in Shakespeare's
## play Love's Labour's Lost, and
##  means "with honorablenesses".
}
\author{
Xiaobei Zhao
}

