% Generated by roxygen2 (4.0.1): do not edit by hand
\name{srep}
\alias{srep}
\title{Replicate and concatenate a string}
\usage{
srep(x, ...)
}
\arguments{
\item{x}{See \code{rep}}

\item{...}{See \code{rep}}
}
\value{
character
}
\description{
Replicate and concatenate a string
}
\examples{
srep("*",5)
}
\author{
Xiaobei Zhao
}

