% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stampme}
\alias{stampme}
\title{Print a message with a time stamp}
\usage{
stampme(x)
}
\arguments{
\item{x}{ANY, an R object.}
}
\description{
Print a message with a time stamp
}
\examples{
stampme('Hello World!')
}
\author{
Xiaobei Zhao
}

