% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stampmsg}
\alias{stampmsg}
\title{Generate a diagnostic message from its arguments,
with timestamp}
\usage{
stampmsg(..., domain = NULL, appendLF = TRUE)
}
\arguments{
\item{...}{see \code{message}}

\item{domain}{see \code{message}}

\item{appendLF}{see \code{message}}
}
\description{
Generate a diagnostic message from its arguments, with timestamp
}
\examples{
stampmsg(LETTERS)
}
\author{
Xiaobei Zhao
}

