% Generated by roxygen2 (4.0.1): do not edit by hand
\name{startswith}
\alias{startswith}
\title{Determine if a character string "starts with" specified characters}
\usage{
startswith(x, char, ignore.case = FALSE)
}
\arguments{
\item{x}{character, a string.}

\item{char}{character to match.}

\item{ignore.case}{logical, whether case is ignored}
}
\value{
logical
}
\description{
Determine if a character string "starts with" specified characters. A modified version of gdata::startsWith.
}
\examples{
startswith('Hello World','hello',ignore.case=TRUE)
}
\author{
Xiaobei Zhao
}

