% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strip}
\alias{strip}
\title{Strip a string with given chars at both ends}
\usage{
strip(x, char = " ")
}
\arguments{
\item{x}{character, a string.}

\item{char}{character to trim.}
}
\value{
character
}
\description{
Strip a string with given chars at both ends
}
\author{
Xiaobei Zhao
}

