% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strsplit.first}
\alias{strsplit.first}
\title{Split a string at the first `split'}
\usage{
strsplit.first(x, split, ...)
}
\arguments{
\item{x}{character, a string to split.}

\item{split,}{see \code{strsplit}}

\item{...,}{see \code{strsplit}}
}
\value{
list
}
\description{
Split a string at the first `split'
}
\examples{
strsplit.first('inFpath="a=1.b=2.c=TRUE"',split="=")
}
\author{
Xiaobei Zhao
}

