% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{UnitTest-class}
\alias{UnitTest}
\alias{UnitTest-class}
\title{Unit testing for developing R packages}
\description{
Unit testing for developing R packages
}
\details{
Unit testing for developing R packages
}
\section{Fields}{

\describe{
\item{\code{pkg}}{character the name of the package}

\item{\code{testDpath}}{character the absolute directory names where to
look for test files. Default: <pkg>/tests}

\item{\code{testFnameRegexp}}{character Regular expression for matching
test file names. Default: *.R}

\item{\code{testFuncRegexp}}{character Regular expression for matching
test functions. Default: test.*}
}}
\examples{
\dontrun{
pkg <- 'Xmisc'
test.obj <- UnitTest$new(pkg=pkg)
test.obj$runme()
}
}
\author{
Xiaobei Zhao
}

