% Generated by roxygen2 (4.0.1): do not edit by hand
\name{valid.mode}
\alias{valid.mode}
\title{Return a valid mode given digits}
\usage{
valid.mode(mode, digits = 4)
}
\arguments{
\item{mode}{character, the mode of the path, see \code{dir.create}.}

\item{digits}{numeric, either 3 or 4.}
}
\value{
mode
}
\description{
Return a valid mode given digits
}
\examples{
valid.mode("777",4)
valid.mode("0777",3)
}
\author{
Xiaobei Zhao
}

