% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vconcat}
\alias{vconcat}
\title{Concatenate vector into a string}
\usage{
vconcat(x, sep = ", ", capsule = FALSE, quote = FALSE)
}
\arguments{
\item{x}{vector}

\item{sep}{character, a delimiter}

\item{capsule}{logical, weather to capsule with `c()'}

\item{quote}{logical, weather to surround elements by double quotes.}
}
\value{
vector
}
\description{
Concatenate vector into a string
}
\examples{
cat(vconcat(head(letters),capsule=TRUE,quote=TRUE),'\\n')
## c("a", "b", "c", "d", "e", "f")

cat(vconcat(head(letters),sep='-'),'\\n')
## a-b-c-d-e-f
}
\author{
Xiaobei Zhao
}

