% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.data.table}
\alias{write.data.table}
\title{A wrapper of write.table}
\usage{
write.data.table(outFpath = "", x, append = FALSE, sep = "\\t",
  quote = FALSE, row.names = FALSE, col.names = !append, logger = NULL,
  ...)
}
\arguments{
\item{outFpath}{file, see \code{write.table}}

\item{x}{see \code{write.table}}

\item{append}{see \code{write.table}}

\item{sep}{see \code{write.table}}

\item{quote}{see \code{write.table}}

\item{row.names}{see \code{write.table}}

\item{col.names}{see \code{write.table}}

\item{logger}{see \code{\link{logme}}}

\item{...}{further arguments passed to `write.table`. See \code{write.table}}
}
\description{
A wrapper of write.table with customized parameters and parsing
}
\author{
Xiaobei Zhao
}

