% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfchunk}
\alias{dfchunk}
\title{Chunk data.frame into parts}
\usage{
dfchunk(x, n)
}
\arguments{
\item{x}{data.frame or matrix}

\item{n}{numeric, the number of chunks}
}
\value{
a list of data.frame
}
\description{
Chunk data.frame into parts
}
\examples{
dfchunk(iris,n=5)
dfchunk(iris[1:5,],n=5)
}
\author{
Xiaobei Zhao
}

