% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ValueParser-class}
\alias{ValueParser}
\alias{ValueParser-class}
\title{Parser for values}
\description{
Parser for values
}
\section{Fields}{

\describe{
\item{\code{value}}{ANY}

\item{\code{type}}{character}
}}
\examples{
ValueParser$new(value="",type='character')$get_value()
ValueParser$new(type='character')$get_value()
}
\author{
Xiaobei Zhao
}

