% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Xmisc-package}
\alias{Xmisc-package}
\title{Xiaobei's miscellaneous classes and functions}
\description{
This is Xiaobei's miscellaneous classes and functions useful when
developing R packages, particularly for OOP using R Reference Class.
}
\details{
Listed are a few highlighted features that may be of common interests.
\itemize{
\item \code{\link{xRefClass-class}}
\item \code{\link{ArgumentParser-class}}
\item \code{\link{logme}}
\item \code{\link{lprintf}}
}
}
\author{
Xiaobei Zhao
}
\references{
Xiaobei Zhao (2014)
}
\keyword{package}

